Move filters to BaseContactTransformer
