+++ src/api/transformers/ContactTransformer.ts
@@ -1,20 +1,16 @@
-import { ContactFilterName, contactFilters } from "@beabee/beabee-common";
 import { TransformPlainToInstance } from "class-transformer";
-import { Brackets, SelectQueryBuilder } from "typeorm";
+import { SelectQueryBuilder } from "typeorm";
 
 import { createQueryBuilder } from "@core/database";
 import PaymentService from "@core/services/PaymentService";
 
 import Contact from "@models/Contact";
 import ContactRole from "@models/ContactRole";
-import ContactProfile from "@models/ContactProfile";
-import PaymentData from "@models/PaymentData";
 import {
   GetContactDto,
   GetContactOptsDto,
   ListContactsDto
 } from "@api/dto/ContactDto";
-import { BaseTransformer } from "@api/transformers/BaseTransformer";
 import ContactRoleTransformer from "@api/transformers/ContactRoleTransformer";
 import ContactProfileTransformer from "@api/transformers/ContactProfileTransformer";
 import { mergeRules } from "@api/utils/rules";
@@ -22,19 +18,12 @@ import { mergeRules } from "@api/utils/rules";
 import { GetContactWith } from "@enums/get-contact-with";
 
 import { AuthInfo } from "@type/auth-info";
-import { FilterHandler, FilterHandlers } from "@type/filter-handlers";
+import { BaseContactTransformer } from "./BaseContactTransformer";
 
-class ContactTransformer extends BaseTransformer<
-  Contact,
+class ContactTransformer extends BaseContactTransformer<
   GetContactDto,
-  ContactFilterName,
   GetContactOptsDto
 > {
-  protected model = Contact;
-  protected filters = contactFilters;
-  // TODO: make protected
-  filterHandlers = contactFilterHandlers;
-
   @TransformPlainToInstance(GetContactDto)
   convert(
     contact: Contact,
@@ -154,87 +143,6 @@ class ContactTransformer extends BaseTransformer<
     }
   }
 }
-
-// Field handlers
-
-function membershipField(field: keyof ContactRole): FilterHandler {
-  return (qb, args) => {
-    const subQb = createQueryBuilder()
-      .subQuery()
-      .select(`cr.contactId`)
-      .from(ContactRole, "cr")
-      .where(`cr.type = 'member'`)
-      .andWhere(args.whereFn(`cr.${field}`));
-
-    qb.where(`${args.fieldPrefix}id IN ${subQb.getQuery()}`);
-  };
-}
-
-function profileField(field: keyof ContactProfile): FilterHandler {
-  return (qb, args) => {
-    const subQb = createQueryBuilder()
-      .subQuery()
-      .select(`profile.contactId`)
-      .from(ContactProfile, "profile")
-      .where(args.whereFn(`profile.${field}`));
-
-    qb.where(`${args.fieldPrefix}id IN ${subQb.getQuery()}`);
-  };
-}
-
-const activePermission: FilterHandler = (qb, args) => {
-  const roleType = args.field === "activeMembership" ? "member" : args.value[0];
-
-  const isIn =
-    args.field === "activeMembership"
-      ? (args.value[0] as boolean)
-      : args.operator === "equal";
-
-  const subQb = createQueryBuilder()
-    .subQuery()
-    .select(`cr.contactId`)
-    .from(ContactRole, "cr")
-    .where(`cr.type = '${roleType}'`)
-    .andWhere(`cr.dateAdded <= :now`)
-    .andWhere(
-      new Brackets((qb) => {
-        qb.where(`cr.dateExpires IS NULL`).orWhere(`cr.dateExpires > :now`);
-      })
-    );
-
-  if (isIn) {
-    qb.where(`${args.fieldPrefix}id IN ${subQb.getQuery()}`);
-  } else {
-    qb.where(`${args.fieldPrefix}id NOT IN ${subQb.getQuery()}`);
-  }
-};
-
-function paymentDataField(field: string): FilterHandler {
-  return (qb, args) => {
-    const subQb = createQueryBuilder()
-      .subQuery()
-      .select(`pd.contactId`)
-      .from(PaymentData, "pd")
-      .where(args.whereFn(`pd.${field}`));
-
-    qb.where(`${args.fieldPrefix}id IN ${subQb.getQuery()}`);
-  };
-}
-
-const contactFilterHandlers: FilterHandlers<ContactFilterName> = {
-  deliveryOptIn: profileField("deliveryOptIn"),
-  newsletterStatus: profileField("newsletterStatus"),
-  tags: profileField("tags"),
-  activePermission,
-  activeMembership: activePermission,
-  membershipStarts: membershipField("dateAdded"),
-  membershipExpires: membershipField("dateExpires"),
-  contributionCancelled: paymentDataField("cancelledAt"),
-  manualPaymentSource: (qb, args) => {
-    paymentDataField("data ->> 'source'")(qb, args);
-    qb.andWhere(`${args.fieldPrefix}contributionType = 'Manual'`);
-  }
-};
 
 export async function loadContactRoles(contacts: Contact[]): Promise<void> {
   if (contacts.length > 0) {
